requirements.txt
pillow
urllib3
discord.py
asyncpg
asyncpraw
aiofiles
openai
websockets